requirements.txt
pytest
sphinx
sphinx-rtd-theme
pillow